+++ src/mkdocs_ezglossary_plugin/glossary.py
@@ -1,6 +1,5 @@
 import hashlib
 import logging
-import html
 
 log = logging.getLogger(__name__)
 
@@ -12,7 +11,7 @@ def get_id(section: str, term: str, linktype: str, n: str):
 class Entry:
     """ An entry in the glossary. """
 
-    def __init__(self, target, page, definition):
+    def __init__(self, target, page, definition, section, term):
         self.target = target
         """ The anchor to directly point to this specific link. """
 
@@ -22,6 +21,12 @@ class Entry:
         self.definition = definition
         """ The definition of the term. """
 
+        self.section = section
+        """ The section of the term of this entry. (Since `v1.7.0a1`) """
+
+        self.term = term
+        """ The term of this entry. (Since v1.7.0a1) """
+
     def __repr__(self):
         return self.page.url
 
@@ -29,16 +34,17 @@ class Entry:
 class Glossary:
     """ The complete glossary for all sections """
 
-    def __init__(self):
+    def __init__(self, ignore_case):
         self.clear()
+        self.ignore_case = ignore_case
 
     def add(self, section, term, linktype, page, definition=None, anchor=None):
-        term = html.unescape(term.strip())
+        term = term.strip()
         log.debug(f"glossary.add({section}, {term}, {linktype}, '{definition}', {anchor})")
         links = self._links(section, term, linktype)
         _id = get_id(section, term, linktype, len(links))
         anchor = _id if anchor is None else anchor
-        links[_id] = Entry(anchor, page, definition)
+        links[_id] = Entry(anchor, page, definition, section, term)
         return _id
 
     def has(self, section: str) -> bool:
@@ -74,7 +80,7 @@ class Glossary:
             Returns:
                 A list of either `definitions` or `references` for the given term.
         """
-        links = self._links(section, html.unescape(term), linktype).values()
+        links = self._links(section, term, linktype).values()
         return list(links)
 
     def definition(self, section: str, term: str) -> str:
@@ -89,7 +95,7 @@ class Glossary:
             Returns:
                 The definition of the term
         """
-        defs = self.get(section, html.unescape(term), 'defs')
+        defs = self.get(section, term, 'defs')
         if not defs:
             return ""
         return defs[0].definition
@@ -116,8 +122,13 @@ class Glossary:
 
     def _term(self, section, term):
         _section = self._section(section)
-        if term not in _section:
-            self._section(section)[term] = {'defs': {}, 'refs': {}}
+        if term in _section:
+            return self._section(section)[term]
+        if self.ignore_case:
+            for candidate in _section.keys():
+                if candidate.upper() == term.upper():
+                    return self._section(section)[candidate]
+        self._section(section)[term] = {'defs': {}, 'refs': {}}
         return self._section(section)[term]
 
     def _section(self, section):