+++ tests/test_definitions.py
@@ -31,7 +31,7 @@ def test_default_config(simple, config):
     assert len(html.xpath(str(dl))) == 0
 
 
-def test_inline_refs(simple, config):
+def test_inline_refs_short(simple, config):
     config['inline_refs'] = "short"
     html = mock.render_single(simple, config)
 
@@ -45,6 +45,18 @@ def test_inline_refs(simple, config):
     dl = dl.has(xpath.dd(text="*third term").has(xpath.a(title="Hello",
                                                          href="../simple.md#" + get_id("test", "third", "refs", 0),
                                                          text="*[1]")))
+    assert len(html.xpath(str(dl))) == 1
+
+
+def test_inline_refs_list(simple, config):
+    config['inline_refs'] = "list"
+    html = mock.render_single(simple, config)
+
+    dl = xpath.dl()
+    dl = dl.has(xpath.dt.has(xpath.a(id=get_id("test", "third", "defs", 0), text="third")))
+    dl = dl.has(xpath.dd(text="*third term").has(xpath.a(title="Hello",
+                                                         href="../simple.md#" + get_id("test", "third", "refs", 0),
+                                                         text="*Hello")))
     assert len(html.xpath(str(dl))) == 1